+++ app/controllers/concerns/index_actions.rb
@@ -1,7 +1,7 @@
 module IndexActions
   ## routed actions
   def list
-    render json: create_datatable
+    render json: create_datatable.serverside
   end
   def index
     table = create_datatable
@@ -10,23 +10,21 @@ module IndexActions
     respond_to do |format|
       format.html {
         render :index, locals: {
-          table: table.ajax(serverside: true, url: url_for(action: :list, format: :json, params: params.permit!))
+          table: table.ajax(serverside: true, url: url_for(action: :list, format: :json, params: params.permit!)).defer_load
         }
       }
-      format.json { render json:
-        table.data.limit(max_limit).map do |item|
-          column_names.map do |column|
-            [column, item.send(column)]
-          end.to_h
-        end
+      format.json {
+        #render json: table.expand_data(table.data.limit(max_limit))
+        render json: table.as_json
       }
       format.csv {
         require 'csv'
         csv = CSV.generate(headers: table.column_names, write_headers: true) do |csv|
-          table.data.limit(max_limit).each do |row|
-            csv << table.column_names.map {|k| row.send(k)}
+          table.as_json.each do |row|
+            csv << row
           end
         end
+        
         send_data csv, filename: "#{controller_name}.csv" }
     end
   end