+++ app/datatables/datatable/defer_loadable.rb
@@ -1,0 +1,18 @@
+module Datatable::DeferLoadable
+  def defer_load
+    self.extend Datatable::DeferLoading
+    self
+  end
+end
+module Datatable::DeferLoading
+  def settings
+    super.merge(deferLoading: records.count)
+  end
+  def manipulate(r)
+    super(r)
+      .order(default_orders.map {|column, dir| [column_defs[column].source, dir].join(' ')})
+      .limit(settings[:pageLength] || 25).decorate
+
+
+  end
+end