+++ app/datatables/scores_datatable.rb
@@ -1,33 +1,30 @@
 class ScoresDatatable < IndexDatatable
   def initialize(*args)
     super *args
-    self.hidden_columns = [:competition_type, :competition_class, :competition_name,
-                          :season, ]
-    sources.update(
-                   name: "scores.name",
-                   category: "scores.category",
-                   competition_name: "competitions.name",
-                   competition_class: "competitions.competition_class",
-                   competition_type: "competitions.competition_type",
-                   season: "competitions.season",
-                   skater_name: "skaters.name",
-                   nation: "skaters.nation",
-                   )
-
-    add_filters(:name, :competition_name, :competition_class, :competition_type, :skater_name, operator: :matches)
-    add_filters(:category, :segment, :nation, :season)
-  end
-  def fetch_records
-    Score.includes(:competition, :skater).references(:competition, :skater).all
-  end
-  def columns
-    [:name, :competition_name, :competition_class, :competition_type,
+    
+    self.columns = [:name, :competition_name, :competition_class, :competition_type,
      :category, :segment, :season, :date,
      :result_pdf, :ranking, :skater_name, :nation,
      :tss, :tes, :pcs, :deductions, :base_value
     ]
+
+    column_defs.sources = {
+      name: "scores.name",
+      category: "scores.category",
+      competition_name: "competitions.name",
+      competition_class: "competitions.competition_class",
+      competition_type: "competitions.competition_type",
+      season: "competitions.season",
+      skater_name: "skaters.name",
+      nation: "skaters.nation",
+    }
+
+    [:competition_type, :competition_class, :competition_name, :season, ].each do |column|
+      self.column_defs[column].visible = false
+    end
+    self.default_orders = [[:date, :desc]]
   end
-  def default_orders
-    [[:date, :desc]]
+  def fetch_records
+    Score.includes(:competition, :skater).references(:competition, :skater).all
   end
 end