+++ app/models/application_record.rb
@@ -6,12 +6,5 @@ class ApplicationRecord < ActiveRecord::Base
       @_list_cache ||= {}
       @_list_cache[key] ||= distinct.pluck(key).compact
     end
-
-    def arel_table_by_operator(key, operator_str, value)
-      operators = {'=' => :eq, '>' => :gt, '>=' => :gteq,
-        '<' => :lt, '<=' => :lteq}
-      operator = operators[operator_str] || :eq
-      arel_table[key].send(operator, value.to_f)
-    end
   end
 end