+++ app/models/element.rb
@@ -1,35 +1,8 @@
 class Element < ApplicationRecord
+  include ScoreVirtualAttributes
+  
   ## relations
   belongs_to :score
-
-  ## for datatable
-  def score_name
-    score.name
-  end
-  def competition_name
-    score.competition.name
-  end
-  def category
-    score.category
-  end
-  def segment
-    score.segment
-  end
-  def date
-    score.competition.start_date
-  end
-  def season
-    score.competition.season
-  end
-  def ranking
-    score.ranking
-  end
-  def skater_name
-    score.skater.name
-  end
-  def nation
-    score.skater.nation
-  end
 
   ## scopes
   scope :recent, ->{ joins(:score).order("scores.date desc") }